BUG-119067: HWC: For SaveMode overwrite, introduce truncate option

+++ src/main/java/com/hortonworks/spark/sql/hive/llap/HiveWarehouseDataSourceWriter.java
@@ -145,6 +145,7 @@ public class HiveWarehouseDataSourceWriter implements DataSourceWriter {
   private void handleWriteWithSaveMode(String database, String table, Connection conn) {
     boolean createTable = false;
     boolean loadData = false;
+    String truncateOption = options.get(HWConf.TRUNCATE_OPTION_KEY);
     switch (mode) {
       case ErrorIfExists:
         if (tableExists) {
@@ -163,9 +164,16 @@ public class HiveWarehouseDataSourceWriter implements DataSourceWriter {
         break;
       case Overwrite:
         if (tableExists) {
-          DefaultJDBCWrapper.dropTable(conn, database, table, false);
+          //option("truncate", "true")
+          if ("true".equalsIgnoreCase(truncateOption)) {
+            DefaultJDBCWrapper.truncateTable(conn, database, table);
+          } else {
+            DefaultJDBCWrapper.dropTable(conn, database, table, false);
+            createTable = true;
+          }
+        } else {
+          createTable = true;
         }
-        createTable = true;
         loadData = true;
         break;
       case Ignore:
@@ -177,8 +185,8 @@ public class HiveWarehouseDataSourceWriter implements DataSourceWriter {
         break;
     }
 
-    LOG.info("Handling write: database:{}, table:{}, savemode: {}, tableExists:{}, createTable:{}, loadData:{}",
-        database, table, mode, tableExists, createTable, loadData);
+    LOG.info("Handling write: database:{}, table:{}, savemode: {}, tableExists:{}, createTable:{}, loadData:{}, truncate:{}",
+        database, table, mode, tableExists, createTable, loadData, truncateOption);
     if (createTable) {
       String createTableQuery = SchemaUtil.buildHiveCreateTableQueryFromSparkDFSchema(schema, database, table);
       DefaultJDBCWrapper.executeUpdate(conn, database, createTableQuery, true);
